requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.2
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.2.3
